# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: reVoAgent API
  description: Enterprise AI Development Platform with Cost-Optimized Model Management
  version: 1.0.0
  contact:
    name: reVoAgent Support
    url: https://github.com/heinzdev11/reVoAgent
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.revoagent.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/agents/{agentId}:
    get:
      summary: Get agent details
      tags: [Agents]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          example: code-analyst
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetails'
  /api/agents/{agentId}/tasks:
    post:
      summary: Start agent task
      tags: [Agents]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          example: code-analyst
        name:
          type: string
          example: Code Analyst
        status:
          type: string
          enum: [active, inactive, busy]
        specialization:
          type: string
        performance_score:
          type: number
          format: float
    AgentDetails:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            capabilities:
              type: array
              items:
                type: string
            performance_metrics:
              type: object
            configuration:
              type: object
    TaskRequest:
      type: object
      required:
        - task_type
        - input
      properties:
        task_type:
          type: string
          example: code_analysis
        input:
          type: object
        options:
          type: object
    TaskResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [started, running, completed, failed]
        estimated_completion:
          type: string
        agent_id:
          type: string
...